Moved all of the proxy classes into 'codegen/proxy' #690
