+++ src/codegen/cc_hash_table.cpp
@@ -12,7 +12,7 @@
 
 #include "codegen/cc_hash_table.h"
 
-#include "codegen/cc_hash_table_proxy.h"
+#include "codegen/proxy/cc_hash_table_proxy.h"
 #include "codegen/hash.h"
 #include "codegen/lang/if.h"
 #include "codegen/lang/loop.h"
