+++ src/codegen/function_builder.cpp
@@ -12,7 +12,7 @@
 
 #include "codegen/function_builder.h"
 
-#include "codegen/runtime_functions_proxy.h"
+#include "codegen/proxy/runtime_functions_proxy.h"
 #include "common/exception.h"
 
 namespace peloton {
