+++ src/codegen/operator/hash_join_translator.cpp
@@ -10,10 +10,10 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "include/codegen/operator/hash_join_translator.h"
+#include "codegen/operator/hash_join_translator.h"
 
 #include "codegen/lang/if.h"
-#include "codegen/oa_hash_table_proxy.h"
+#include "codegen/proxy/oa_hash_table_proxy.h"
 #include "codegen/expression/tuple_value_translator.h"
 #include "codegen/lang/vectorized_loop.h"
 #include "expression/tuple_value_expression.h"
