+++ src/codegen/proxy/transaction_runtime_proxy.cpp
@@ -10,12 +10,12 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "codegen/transaction_runtime_proxy.h"
+#include "codegen/proxy/transaction_runtime_proxy.h"
 
-#include "codegen/executor_context_proxy.h"
-#include "codegen/data_table_proxy.h"
-#include "codegen/tile_group_proxy.h"
-#include "codegen/transaction_proxy.h"
+#include "codegen/proxy/executor_context_proxy.h"
+#include "codegen/proxy/data_table_proxy.h"
+#include "codegen/proxy/tile_group_proxy.h"
+#include "codegen/proxy/transaction_proxy.h"
 #include "codegen/value_proxy.h"
 
 namespace peloton {
