+++ src/codegen/tile_group.cpp
@@ -15,9 +15,9 @@
 #include "catalog/schema.h"
 #include "codegen/lang/if.h"
 #include "codegen/lang/loop.h"
-#include "codegen/runtime_functions_proxy.h"
+#include "codegen/proxy/runtime_functions_proxy.h"
 #include "codegen/scan_callback.h"
-#include "codegen/tile_group_proxy.h"
+#include "codegen/proxy/tile_group_proxy.h"
 #include "codegen/varlen.h"
 #include "codegen/vector.h"
 #include "codegen/lang/vectorized_loop.h"
