+++ src/codegen/type/tinyint_type.cpp
@@ -14,7 +14,7 @@
 
 #include "codegen/lang/if.h"
 #include "codegen/value.h"
-#include "codegen/values_runtime_proxy.h"
+#include "codegen/proxy/values_runtime_proxy.h"
 #include "codegen/type/boolean_type.h"
 #include "codegen/type/integer_type.h"
 #include "common/exception.h"
