+++ src/codegen/type/varchar_type.cpp
@@ -13,7 +13,7 @@
 #include "codegen/type/varchar_type.h"
 
 #include "codegen/value.h"
-#include "codegen/values_runtime_proxy.h"
+#include "codegen/proxy/values_runtime_proxy.h"
 #include "codegen/type/boolean_type.h"
 #include "codegen/type/integer_type.h"
 #include "common/exception.h"
